backend/requirements.txt
fastapi
uvicorn
PyPDF2
docx2txt
google-genai
pydantic
python-dotenv
sqlalchemy
